Add Feat: Authentication and security

+++ service/ProductService.js
@@ -1,11 +1,23 @@
+import { getToken } from "../helpers/get-token.js";
+import { getUserByToken } from "../helpers/get-user-by-token.js";
+
 class ProductService {
-    constructor(productRepository, batchRepository) {
+    constructor(productRepository, batchRepository, userRepository) {
         this.productRepository = productRepository;
         this.batchRepository = batchRepository;
+        this.userRepository = userRepository;
     };
 
-    async showAllProducts() {
-        const products = await this.productRepository.findAll();
+    async showAllProducts(data) {
+
+        //Get the Token
+        const token = getToken(data);
+
+        //Get User by Token
+        const user = await getUserByToken(token)
+
+        //Get Products on User
+        const products = await this.productRepository.findAllByPk({ where: { UserId: user.id } });
 
         const plainProducts = products.map(product => product.get({ plain: true }));
 
@@ -17,23 +29,29 @@ class ProductService {
             currency: 'BRL'
         });
 
-        const data = {
+        const productsInfo = {
             products: plainProducts,
             totalPrice
         }
 
-        return data;
+        return productsInfo;
     };
 
     async createProduct(data) {
-        const name = data.name;
-        const category = data.category;
-        const unit_price = data.unit_price;
+        const name = data.body.name;
+        const category = data.body.category;
+        const unit_price = data.body.unit_price;
+
+
+        //Get Owner Product
+        const token = getToken(data);
+        const user = await getUserByToken(token);
 
         const products = {
             name,
             category,
-            unit_price
+            unit_price,
+            UserId: user.id
         }
 
         const product = await this.productRepository.create(products);
@@ -42,13 +60,24 @@ class ProductService {
     };
 
     async deleteProduct(data) {
-        const id = data.id;
-        
-        //Remove all batches associates the Product ID
-        await this.batchRepository.destroy(id);
+        const id = data.params.id;
 
-        //After the program excludes the Product
+        //Get the Token
+        const token = getToken(data);
+
+        //Get User by Token
+        const user = await getUserByToken(token)
+
+        const productFromDB = await this.productRepository.findByPk(id);
+
+        if (user.id !== productFromDB.UserId) {
+            return "Esse produto não pertence a você.";
+        }
+
+        await this.batchRepository.destroy(id);
         await this.productRepository.destroy(id);
+
+        return null; 
     }
 };
